requirements.txt
python-telegram-bot==13.7
pyrebase4==4.5.0
web3==5.24.0
python-dotenv==0.19.0
requests==2.26.0